renamed DefaultVehiclesRepository & its interface
